Implement better error handling for iOS
